--- src/components/Pages/home.module.css
@@ -1,134 +1,0 @@
-.homeContainer{
-    display: flex;
-    flex-direction: row;
-    padding: 10px;
-    position: relative;
-
-    .leftSide {
-        background: linear-gradient(219.62deg, rgba(109, 84, 235, 0.6) 28.31%, rgba(101, 35, 146, 0.6) 66.76%);
-
-        .navLinks {
-            
-            padding: 40px 339px 28px 16px;
-        }
-        .balanceContainer {
-            display: flex;
-            flex-direction: column;
-            justify-content: center;
-            gap: 12px;
-            padding: 8px 249px 11px 56px;
-
-            box-shadow: 1px 9px 15px 0px #00000033;
-            background-color: #523B7E99;
-            width: 480x;
-            height: 80px;
-            font-family: Poppins;
-            font-size: 12px;
-            font-weight: 400;
-            line-height: 18px;
-            text-align: left;
-            color: #FFFFFF66;
-
-
-                
-            .balanceSold {
-                font-family: Poppins;
-                font-size: 30px;
-                font-weight: 700;
-                line-height: 45px;
-                text-align: left;
-                text-align: left;
-                color: white;
-            }   
-  
-        }
-
-        .currencyContainer {
-            width: 480px;
-            height: 380px;
-            margin-top: 32px;
-            box-shadow: 0px 4px 60px 0px #00000040;
-            text-align: center; 
-        }
-
-    }
-   
-    .rightSide {
-        background: linear-gradient(219.62deg, rgba(109, 84, 235, 0.6) 28.31%, rgba(101, 35, 146, 0.6) 66.76%);
-        width: 100%;
-        
-        
-        .statisticContainer {
-            display: flex;
-            flex-direction: column;
-            justify-content: center;
-            gap: 12px;
-            padding: 8px 249px 11px 56px;
-            position: relative;
-           
-            
-            .statistics {
-                width: 100%;
-                height: 100%;
-              
-
-                .table {
-                    gap: 0px;
-                    border: 1px 0px 0px 0px;
-                    opacity: 0px;
-                    width: 100%;
-                    height: 100%;
-                    padding: 0px 16px 20px 16px;
-                    box-shadow: 0px 4px 60px 0px #00000040;
-                    font-family: Poppins;
-                    
-                   thead{
-                    font-size: 16px;
-                    font-weight: 600;
-                    line-height: 24px;
-                    text-align: left;
-                    color: #FFFFFF;
-                   }
-                   tbody {
-                    width: 565px;
-                    height: 21px;
-                    top: 198px;
-                    left: 569px;
-                    gap: 0px;
-                    opacity: 0px;
-                    border: 1px 0px 0px 0px;
-                    color: #FFFFFF;
-                   }
-                    
-                }
-            }
-
-
-
-            .btnContainer {
-                position: absolute;
-                bottom: -570px; 
-                right: 10px; 
-            
-
-                .addBtn {
-                    width: 44px;
-                    height: 44px;  
-                    background: linear-gradient(96.76deg, #FFC727 -16.42%, #9E40BA 97.04%, #7000FF 150.71%);
-                    border-radius: 22px;
-                    box-shadow: 0px 4px 60px 0px #00000040;
-                    border: 0px;
-                    color: white;
-                    font-size: 20px;
-                    margin-right: 10px;
-                  }
-            }
-         
-        }
-
-    }
-
-  
-}
-
-
