+++ src/components/pages/Statistics.jsx
