+++ src/components/pages/dashboard/Balance.jsx
@@ -1,0 +1,12 @@
+import styles from './home/home.module.css';
+
+function Balance() {
+  return (
+    <div className={styles.balanceContainer}>
+        <p>YOUR BALANCE</p>
+        <h2 className={styles.balanceSold}>$2,345</h2>
+    </div>
+  );
+}
+
+export default Balance;
