+++ src/components/pages/dashboard/Currency.jsx
@@ -1,0 +1,11 @@
+import styles from './home/home.module.css';
+
+function Currency() {
+    return (
+        <div className={styles.currencyContainer}>
+            <p>API</p>
+        </div>
+    );
+}
+
+export default Currency;
