+++ src/components/pages/dashboard/Header.jsx
@@ -1,0 +1,9 @@
+function Header() {
+    return (
+        <div>
+            Header
+        </div>
+    );
+}
+
+export default Header;
